Fix the case where market list is not cached, even though the cache is empty.
After a fiat curency change, CacheOrchestrator clears the market cache, but the MarketRepositoryImpl doesn't know that cache is cleared, and relies on it's _lastCacheDate and _cacheTTL variables to decide on cache or not.

Fixed this issue by, adding a isEmpty function on ICacheService interface, implemented it on HiveCacheServer and created a function on local data source of Market feature.

Added a isEmpty check to CacheOrchestrator too so it doesn't clear a cache that is already empty.