Mitigate from ExchangeRate model to DailyExchangeRate model.
The problem with ExchangeRate model was that, it would've caused problems with offline support on some cases.
Now, every exchange rate in FrankfurterApi get's fetched and stored in ExchangeRateModel, and data structure is just like Frankfurter's response json.
When user need's today's exchange rate, only the supported fiat's in FiatCurrency enum will be fetched.
Although, every rate in response will be fetched, only supported currencies will be in the entity, and fetching and caching every currency in response, costs too little in memory, that It would benefit, so much in case the app support's more fiat currencies in the future.